+++ src/Sulu/Bundle/MediaBundle/Media/Manager/MediaManager.php
@@ -813,6 +813,8 @@ class MediaManager implements MediaManagerInterface
                     $fileVersion->getName(),
                     $fileVersion->getStorageOptions()
                 );
+
+                $this->storage->remove($fileVersion->getStorageOptions());
             }
         }